requirements.txt
aiogram==3.1.1
python-dotenv==1.0.0
httpx==0.24.1
pydantic==2.5.0